remove wcf from plasma shared dept
